CB-7691 separate AwsLaunchTest and AwsRepairTest
